refactor: move TUI machinery off BaseAgent onto InteractiveTuiAgent (Option A)
The architecture review flagged that BaseAgent was doing double duty: it
provided both a "minimal agent" implementation (used by codex/opencode/
command) and the full TUI input pipeline as opt-in machinery exposed via
``uses_paste_detection_send``/``uses_submission_signal`` hooks. Cleaner
split: BaseAgent is non-TUI-only, InteractiveTuiAgent owns everything
related to paste-detection, ready-indicator wait, and submission signal.

BaseAgent shrinks to:
- send_message: plain `_send_message_simple` (literal text + Enter)
- wait_for_ready_signal: just runs start_action
- _send_tmux_literal_keys, _send_message_simple
- _capture_pane_content, _check_pane_contains (generic pane helpers)
- _preflight_send_message (default no-op), _raise_send_timeout

InteractiveTuiAgent gains everything else:
- TUI_READY_INDICATOR ClassVar; get_tui_ready_indicator returns it
- uses_submission_signal (default True)
- enter_submission_timeout_seconds field
- _SEND_MESSAGE_TIMEOUT_SECONDS / _TUI_READY_TIMEOUT_SECONDS /
  _DEFAULT_ENTER_SUBMISSION_WAIT_FOR_TIMEOUT_SECONDS /
  _SEND_ENTER_NO_SIGNAL_* constants
- _normalize_for_match / _check_paste_content helpers
- Override send_message to use paste-detection
- Override wait_for_ready_signal to wait for TUI banner on creation
- _send_message_with_paste_detection, _wait_for_paste_visible,
  _is_paste_visible, _wait_for_tui_ready
- _send_enter_and_wait dispatcher + the two paths
  (_send_enter_and_wait_for_submission_signal,
  _send_enter_and_poll_for_input_ready),
  _send_enter_keystroke, _send_enter_and_wait_for_signal,
  _get_last_queue_timestamp

Removed entirely: uses_paste_detection_send. Paste-detection is now
implicit in InteractiveTuiAgent (always on); no hook needed.

Knock-on changes:
- PiCodingAgent migrates from BaseAgent to InteractiveTuiAgent. Drops
  uses_paste_detection_send / get_tui_ready_indicator / _send_enter_and_wait
  overrides. Adds TUI_READY_INDICATOR = "pi v" and uses_submission_signal
  returning False (pi has no UserPromptSubmit hook; the prior
  paste-visibility check + scrollback-persistent banner make the
  poll-based path's check resolve immediately).
- BaseHeadlessAgent drops its dead uses_paste_detection_send /
  get_tui_ready_indicator overrides (nothing on the new BaseAgent
  dispatches on them anymore; headless agents reject sends at preflight).
- HeadlessClaude drops the parallel dead overrides that delegated to
  BaseHeadlessAgent's removed methods.
- Tests rebalanced: paste-detection / Enter-and-poll / wait-for-signal
  tests move to tui_agent_test.py; base_agent_test.py keeps only what
  applies to non-TUI BaseAgent. PiCoding plugin_test.py drops the now
  redundant uses_paste_detection / _send_enter_and_wait tests.
